Here is the application that is used to compare all famous sorting methods for arrays of Integer and Float case values
In this project we are testing the average amount of time we need each of sorting algorithms to sort a randomly created array in ascending order. In thisp program I am having the Insertion sort, Binary Insertion sort, Heap sort, Quick sort (with the ability of choosing the pivot), and Shell sort (with the ability of choosing the pattern). This application can generate random arrays or can use the arrays, that would be provided from file. Also we can test different data types, such as Integer and Float, but only for Insertion sort. As to add this program has two modes of being used: simulation(one launch for each algorithm and extra commentaries) and testing mode (multiply running with different arrays for each size for each algorithm).