Here is the application for testing the main graf algorithms
In this application I have implemented two representations of graf in computer memory, such as neighborhood list and incident matrix. We are going to test two types of algorithms: the Minimal Spanning Tree algorithms (Kruskal and Prim), shortest path search algorithms (Bellman Ford and Dijsktr). For those algorithms we are going to test each of our representations of graf to see what would be the time difference of solving them by each algorithm. In this project we are also having two modes of starting our research: Simulation mode (one time for each algorithm for each representation with big amount of comments) and Test mode (multiply testing of each algorithm with each time generating new graf for each representation for our research to be provided in a correct way).